social_network/requirements.txt
Django==1.11.3
django-custom-user==0.7
pytz==2017.2
django-rest-swagger==2.1.2
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
clearbit==0.1.6
django-filter==1.0.4